collector/requirements.txt
pandas
numpy
requests
pykx
ib_insync
websocket-client
schedule
PyYAML                                                                                                                                                                                                                                                                                                                                                                                          
